+++ src/main/java/com/revature/movie/App.java
@@ -1,10 +1,12 @@
 package com.revature.movie;
 
+/**
+ * The main method will run the movies ServerMovie class
+ */
+
 public class App {
     public static void main(String[] args){
         ServerMovie sm = new ServerMovie();
-       // MovieFile mf = new MovieFile;
-        //mf.httpServerMethod();
     }
     public App(){