+++ CHANGELOG.md
@@ -5,3 +5,6 @@
     2-27-2022
 - The Html prints Movie Search for the title. It also prints a search bar and link to a Movie list.
 <HR>
+
+    3-3-2022
+- Add build method that contains getter and setters. Tested the Movie test class and added movie test class. Also attempted to add movie client server. Attempted to add movie jar executable file.